% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fInd_package.R
\name{find_package}
\alias{find_package}
\title{Find package by a keyword in function}
\usage{
find_package(funkey, data = NULL, exact = FALSE, repo = getOption("repos"))
}
\arguments{
\item{funkey}{one character}

\item{data}{NULL or result of loadData() function.}

\item{exact}{logical, whether to match exactly}

\item{repo}{repository}
}
\value{
a datafram contains package and function
}
\description{
Find package by a keyword in function
}
\examples{
\donttest{
find_package('nomogram')
}

}
