% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.inCRAN.R
\name{is.inCRAN}
\alias{is.inCRAN}
\title{Check package in current CRAN}
\usage{
is.inCRAN(package, data = NULL, repo = getOption("repos"))
}
\arguments{
\item{package}{one or more package names}

\item{data}{NULL or result of loadData() function.}

\item{repo}{repository}
}
\value{
logical
}
\description{
Check whether packages in current CRAN
}
\examples{
is.inCRAN(c('do','export'),repo="https://cloud.r-project.org/")
}
