% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodData.R
\name{loadData}
\alias{loadData}
\title{Load Package Information}
\usage{
loadData(repo = getOption("repos"))
}
\arguments{
\item{repo}{CRAN repository}
}
\value{
a list contains archive, current and available package information on
current repository, and bioconductor package information containing
Software, AnnotatonData, ExperimentData and Workflow from the current
release version. see \code{\link[wherepackage]{where}}
}
\description{
Load Package Information
}
