% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{where}
\alias{where}
\title{Search Package Source}
\usage{
where(data, packages)
}
\arguments{
\item{data}{result of functin loadData()}

\item{packages}{one or more package names}
}
\value{
a dataframe contains package name, version, modified time and source.
}
\description{
Search package from CRAN, Archive, Available and Bioconductor.
}
\examples{

\donttest{

d <- loadData()
where(d,c('purrrogress','do'))

# then use command to install
remotes::install_version('purrrogress','0.1.1')
}

}
